fix(tools): router prompt negative few-shots; float timeouts; deterministic timeout test

+++ backend/app/services/tools/base.py
@@ -20,4 +20,4 @@ class Tool:
     trigger_hints: list[str]   # 聊天门控关键词；也是心跳技能的 trigger_keywords 种子
     params_schema: dict        # 参数名 -> 中文说明（用于生成路由 prompt）
     execute: Callable[[dict], Awaitable["ToolResult"]] = field(compare=False, repr=False)
-    timeout: int = 10          # 执行超时秒数（claude CLI 类工具需调大）
+    timeout: float = 10        # 执行超时秒数（claude CLI 类工具需调大）
